requirements.txt
bidict==0.23.1
blinker==1.9.0
click==8.2.1
Flask==3.1.2
Flask-SocketIO==5.5.1
h11==0.16.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==2.3.3
pandas==2.3.3
pillow==11.3.0
python-dateutil==2.9.0.post0
python-engineio==4.12.2
python-socketio==5.13.0
pytz==2025.2
simple-websocket==1.1.0
six==1.17.0
tqdm==4.67.1
tzdata==2025.2
Werkzeug==3.1.3
wsproto==1.2.0
